requirements.txt
selenium == 4.4.3
pytest == 7.1.3
allure-pytest==2.9.45
pytest-xdist==2.5.0